requirements.txt
flask
pyparsing
openai
python-dotenv

